AttemptTrailActivity:
created:
AddFriend: data class that holds the username and a boolean of isSelected

AddFriendAdapter: this is an adapter for the recyclerView in the dialog_add_friends.xml (it holds a list of the users friends and a checkbox for each user)

dialog_add_friends: holds a recyclerView for the users friends and a button to confirm the users choice

dialog_add_user: has a editText section where the user can add a non-trailBlazers name to add to their current party

item_add_friend: layout for the addFriend

item_trail_user: layout for the recyclerView of current Party

NonTrailBlazeUser: data class to hold the name of the non-trailBlazer

UserAdapter: adapter for the list of users in the current party

activity_attempt_trail: updated the layout for when the user unchecks solo adventure

AttemptTrailActivity:
added the recyclerView for the current Party.
added the addNonTrailBlazer and addTrailBlazer buttons.
created a showAddFriendsDialog that opens a dialog and list the users friends, and then the user can click on the check box to add that user to the current party recyclerView
created a addCurrentUsertoParty fucntion that adds the current user to the party
created a showAddUserDialog, that opens the dialog to add non trailblazer's names and adds the name to the party recycler
created a updatePartyMemberCount fucntion that updates the count of the currenct party based on added members
created a savePartyMembers to firestore and deletePartyMembers function so that we could notifiy watchers easily (and delete the party when the user is done with the trail)