updated the seek bar to be able to go up to 100 miles
the fetchParksAndPlacesMarkers now maintains users current zoom level and camera position while displaying all the parks in the state. The markers will be placed on the map without adjusting the view.

yes i know that parkWithinBounds is not being used anymore and that more then just the states nps are being called... but NPS is not Google Maps... and NPS fetches differently then google... and the parkWithinBounds is cutting off over half of the parks for the NPS even if the parks are within the bounds set...